extract optimization to utility
